+++ src/solana/solanaTypes.ts
@@ -1,15 +1,11 @@
 import {
   asArray,
   asCodec,
-  asEither,
   asMaybe,
-  asNull,
   asNumber,
   asObject,
   asOptional,
   asString,
-  asTuple,
-  asUnknown,
   Cleaner
 } from 'cleaners'
 
@@ -32,7 +28,7 @@ export const asSolanaInfoPayload = asObject({
 })
 
 export const asSolanaWalletOtherData = asObject({
-  newestTxid: asMaybe(asString, '')
+  newestTxid: asMaybe(asObject(asString), () => ({}))
 })
 
 export type SolanaWalletOtherData = ReturnType<typeof asSolanaWalletOtherData>
@@ -40,24 +36,6 @@ export type SolanaWalletOtherData = ReturnType<typeof asSolanaWalletOtherData>
 export const asRpcBalance = asObject({
   value: asNumber
 })
-
-export const asRpcSignatureForAddress = asObject({
-  signature: asString,
-  blocktime: asOptional(asNumber),
-  err: asUnknown
-})
-
-export type RpcSignatureForAddress = ReturnType<typeof asRpcSignatureForAddress>
-
-export const asLatestBlockhash = asObject({
-  value: asObject({
-    blockhash: asString
-  })
-})
-
-export const asGetRecentPrioritizationFees = asArray(
-  asObject({ prioritizationFee: asNumber, slot: asNumber })
-)
 
 export type SafeSolanaWalletInfo = ReturnType<typeof asSafeSolanaWalletInfo>
 export const asSafeSolanaWalletInfo = asSafeCommonWalletInfo
@@ -102,9 +80,6 @@ const asRpcResponse = <T>(cleaner: Cleaner<T>): Cleaner<{ result: T }> =>
     result: cleaner
     // id: 1
   })
-export const asRpcSignatureForAddressResponse = asTuple(
-  asRpcResponse(asArray(asRpcSignatureForAddress))
-)
 export const asAccountBalance = asRpcResponse(asRpcBalance)
 export type AccountBalance = ReturnType<typeof asAccountBalance>
 
@@ -118,63 +93,6 @@ const asRpcTokenBalance = asObject({
   })
 })
 export const asTokenBalance = asRpcResponse(asRpcTokenBalance)
-export type TokenBalance = ReturnType<typeof asTokenBalance>
-
-export const asAccountInfo = asObject({
-  // context: { apiVersion: '1.16.19', slot: 232816397 },
-  value: asEither(
-    asNull,
-    asObject({
-      // data: ['', 'base58'],
-      // executable: false,
-      // lamports: 254385770,
-      // owner: '11111111111111111111111111111111',
-      // rentEpoch: 0,
-      // space: 0
-    })
-  )
-})
-
-const asTxTokenBalance = asObject({
-  accountIndex: asNumber,
-  mint: asString,
-  owner: asString,
-  programId: asString,
-  uiTokenAmount: asObject({
-    amount: asString
-    // "decimals": asNumber,
-    // "uiAmount": 7150.402259,
-    // "uiAmountString": "7150.402259"
-  })
-})
-
-export const asRpcGetTransaction = asObject({
-  meta: asObject({
-    err: asOptional(asUnknown),
-    fee: asNumber,
-    innerInstructions: asArray(asUnknown),
-    postBalances: asArray(asNumber),
-    postTokenBalances: asArray(asMaybe(asTxTokenBalance)),
-    preBalances: asArray(asNumber),
-    preTokenBalances: asArray(asMaybe(asTxTokenBalance))
-  }),
-  slot: asNumber,
-  transaction: asObject({
-    message: asObject({
-      accountKeys: asArray(
-        asObject({
-          pubkey: asString
-        })
-      ),
-      instructions: asArray(asUnknown),
-      recentBlockhash: asString
-    }),
-    signatures: asArray(asString)
-  })
-})
-export const asTransaction = asRpcResponse(asRpcGetTransaction)
-
-export type RpcGetTransaction = ReturnType<typeof asTransaction>
 
 export const asBlocktime = asRpcResponse(asNumber)
 export type Blocktime = ReturnType<typeof asBlocktime>
@@ -190,6 +108,7 @@ export const asSolanaInitOptions = asObject({
   alchemyApiKey: asOptional(asString),
   poktPortalApiKey: asOptional(asString)
 })
+export type SolanaInitOptions = ReturnType<typeof asSolanaInitOptions>
 
 export const asSolanaCustomFee = asObject({
   microLamports: asString